# documentation/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/kcb/funds-transfer": {
      "post": {
        "operationId": "KcbController_fundsTransfer",
        "summary": "Initiate a standalone KCB funds transfer",
        "description": "Calls KCB OAuth and the KCB Funds Transfer API. PayAssure forwards the OAuth token authorization header internally and never returns it to the client. This asynchronous integration is standalone and is not invoked by settlement initiation, payment callbacks, or payout dispatch.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "beneficiaryDetails": "JOHN DOE",
                  "companyCode": "KE0010001",
                  "creditAccountNumber": "1279287799",
                  "currency": "KES",
                  "debitAccountNumber": "1279258233",
                  "debitAmount": 26,
                  "paymentDetails": "UT Fund withdrawal",
                  "transactionReference": "FT1234567890",
                  "transactionType": "IF",
                  "beneficiaryBankCode": "01"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "KCB transfer accepted for asynchronous processing.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "provider": "KCB",
                    "transactionReference": "FT1234567890",
                    "response": {
                      "statusCode": "0",
                      "statusMessage": "Success",
                      "statusDescription": "Request received for processing",
                      "merchantID": "263eb626-3fe7-4662-813e-f6f2962219e1",
                      "retrievalRefNumber": "PCI663RSS"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid transfer request."
          },
          "503": {
            "description": "KCB authentication or transfer request failed."
          }
        },
        "tags": [
          "KCB Funds Transfer"
        ]
      }
    },
...
  },
  "info": {
    "title": "PayAssure API",
    "description": "PayAssure settlement and onboarding APIs",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Provide a valid JWT access token in the format: Bearer <access_token>."
      }
    },
...
  }
}